% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-corr_report.R
\name{corr_report}
\alias{corr_report}
\title{Perform a simple bivaraite correlation}
\usage{
corr_report(var1, var2, data, method = "pearson",
  use = "pairwise.complete.obs")
}
\arguments{
\item{var1}{a string indicating the 1st variable you want to correlate}

\item{var2}{a string indicating the 2nd variable you want to correlate}

\item{data}{a \code{data.frame} containing the variables you are correlating}

\item{method}{a string, the same as the method argument that is normally passed to \code{cor}, indicates which
correlation coefficient to compute. Must be "pearson" (default), "kendall", or "spearman".}

\item{use}{a string, the same 'use' arguemnt normally passed to \code{cor}. Indicates what to do with
missing values. Options must be "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs" (default)}
}
\value{
this fucntion returns a list containing the following values:

\describe{
  \item{vars}{variables that were correlated}
  \item{eff}{numeric value representing the correlation coefficient}
  \item{n}{sample size}
  \item{report}{string that provides the correlation coefficient and sample size}
  \item{method}{the method used to compute the correlation coefficient}
}
}
\description{
Perform a simple bivaraite correlation
}
