% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-corr_table.R
\name{corr_table}
\alias{corr_table}
\title{Create a formatted correlation table}
\usage{
corr_table(data, sample_size = T, use = "pairwise",
  method = "pearson", stats = NULL, c.names = names(data),
  change = F, numbered = F, flagged = T)
}
\arguments{
\item{data}{a \code{data.frame} containing only the variable to be included in the table}

\item{sample_size}{logical. TRUE will add the sample size used in each pairwise correlation
and display them in the upper triangle of the table. FALSE will leave the upper triangle empty}

\item{use}{a string, same as the arguemnt used in \code{psych::corr.table}. Indicates how to
calculate correlations with missing data. Defualt is \code{"pairwise"}.}

\item{method}{a string, same as the argument in \code{psych::corr.table}. Indicates the type of correlation
coefficient to compute. Default is \code{"pearson"}}

\item{stats}{defaults to NULL and does not include any descriptive statistics. Otherwise should be a
vector of strings naming the stats that \psych{describe} outputs and will be displayed below the
correlation matrix}

\item{c.names}{a character vector of user-provided names for the names of the variables in the table.}

\item{change}{logical, indicates if you would like to change the column names of the table to \code{c.names}}

\item{numbered}{logical, indicates if you would like to use numbers for the column names}

\item{flagged}{logical, indicates if you would like to flag significant (p <. 05) correlations in the table}
}
\value{
a \code{data.frame} containing the correlation matrix and descriptive statistics
}
\description{
Create a formatted correlation table
}
