% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-beta_report.R
\name{lm_beta_report}
\alias{lm_beta_report}
\title{Perform a simple multiple regression and get back a standardized beta}
\usage{
lm_beta_report(IV, DV, Covariate, data)
}
\arguments{
\item{IV}{a string indicating the name of the independent variable.}

\item{DV}{a string indicating the name of the dependent variable.}

\item{Covariate}{a string indicating the name of a covariate.}

\item{data}{a \code{data.frame} that you want to use for the analysis.}
}
\value{
This function returns a list containing the following items:

\describe{
  \item{vars}{variables that were used in the model}
  \item{model_lm}{the lm object with undstandardized betas}
  \item{model_lm_z}{the lm.beta obeject with standardized betas}
  \item{table}{table of betas and standardized betas with standard error}
  \item{n}{sample size}
  \item{eff}{standarized beta weight for the IV}
  \item{report}{string that provides the beta weight and sample size}
}
}
\description{
Perform a simple multiple regression and get back a standardized beta
}
