% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-lookup_table.R
\name{lookup_table}
\alias{lookup_table}
\title{Create a \code{data.frame} of variable names and labels.}
\usage{
lookup_table(data, ...)
}
\arguments{
\item{data}{a \code{data.frame} with labels that you would like to
convert to a table of variable names and labels}

\item{...}{other arguments that can be passed to \code{paste()}}
}
\value{
a \code{data.frame} with three columns. The first is the names of the variables in
 \code{data}, the second is their labels, and the third are any value labels associated with each variable.
 If a certain variable does not have an variable label or value label it will be filled in with "".
 If there are one or more empty labels, a message will print
 listing the variables that do not have either variable and/or value labels.
}
\description{
Create a \code{data.frame} of variable names and labels.
}
