% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip-row-operations.R
\name{tidy_avg_rows}
\alias{tidy_avg_rows}
\title{Updated version of \code{average_rows()}}
\usage{
tidy_avg_rows(data, selector, reverse = NULL, show = F, mini = NULL,
  maxi = NULL)
}
\arguments{
\item{data}{a \code{data.frame} to use for averaging across rows}

\item{selector}{a \code{dplyr::select} helper function, for example \code{starts_with()}}

\item{reverse}{a numeric vector indicating items or variables to reverse code using
\code{reverse.code} from the \code{psych} package.}

\item{show}{choose whether or not show the results of the reverse coding. If TRUE,
the result of the function will be a \code{data.frame} with the original columns passed to
the function along with the reverse coded columns denoted with an "_r". Useful for checking to
make sure reverse coding was done to the appropriate columns.}

\item{mini}{the minimum value of a scale score (used when reverse = T)}

\item{maxi}{the maximum value of a scale score (used when reverse = T)}
}
\value{
a vector of the averaged values for each row from the specified columns of the data frame.

This function, and those specified below in 'See Also', work best when used
inside a call to \code{dplyr}'s \code{mutate()} for creating new variables.
The point of the function is to make it easy to use select helpers from
the tidyselect (and used in dplyr extensively) to select variables for aggregation,
such as an average score for a number of questionnaire items, the sum of correct responses,
or finding the number of missing values in a row.
}
\description{
Average values across multiple columns for each row of a \code{data.frame}
}
\seealso{
\code{\link{tidy_sum_rows}} for row summing, \code{\link{tidy_sum_missings}} for summing
  missing values, and \code{\link{tidy_sd_rows}} for calculating standard deviations across rows.
}
