% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip-row-operations.R
\name{tidy_sd_rows}
\alias{tidy_sd_rows}
\title{Updated version of \code{sd_rows()}}
\usage{
tidy_sd_rows(data, selector)
}
\arguments{
\item{data}{a \code{data.frame} to use for row standard deviations}

\item{selector}{a \code{dplyr::select} helper function, for example \code{starts_with()}}
}
\value{
a vector of the standard deviations for each row from the specified columns of the data frame.
}
\description{
Updated version of \code{sd_rows()}
}
\section{Details}{

This function, and those specified below in 'See Also', work best when used
inside a call to \code{dplyr}'s \code{mutate()} for creating new variables.
The point of the function is to make it easy to use select helpers from
the tidyselect (and used in dplyr extensively) to select variables for aggregation,
such as an average score for a number of questionnaire items, the sum of correct responses,
or finding the number of missing values in a row.
}

\seealso{
\code{\link{tidy_avg_rows}} for row averaging across rows, \code{\link{tidy_sum_rows}} for row summing
, and \code{\link{tidy_sum_missings}} for summing missing values
}
