% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip-winzor_z.R
\name{winsor_z}
\alias{winsor_z}
\title{Winzorize outliers above a particular standardized value}
\usage{
winsor_z(x, z)
}
\arguments{
\item{x}{a numeric vector of values to winsorize}

\item{z}{a single numeric value indicating the threshold above which
extreme scores should be winsorized.}
}
\value{
a numeric vector with original values with extreme values replaced
 with the next extreme value in the distribution. The number of replacemnets
 and replacement values depend on the \code{z} argument.
}
\description{
Winzorize outliers above a particular standardized value
}
